add facilities_non_primary_staff as a standard role (Recidiviz/recidiviz-data#38384)
## Description of the change

Facilities equivalent of
https://github.com/Recidiviz/recidiviz-data/pull/38059. Requested in
slack:
https://recidiviz.slack.com/archives/C026UPMAX4G/p1739481296042619?thread_ts=1739388701.665239&cid=C026UPMAX4G

@alexabatino for context, we recently made it so that every user in the
admin panel needs to have a "standard role" (or predefined role, they're
the same thing) in addition to any other user-defined roles they might
have, so that we can more correctly identify things like "the set of
line staff users who have access to our tools" for usage tracking. so
now, if there's a role we want to use elsewhere in the codebase or just
makes sense across states, it needs to be in this list.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1595d00c8844400e9b5c82c7d11d990f950bdb47