[Dataflow] Add a gcloudignore file for sandbox pipeline builds (Recidiviz/recidiviz-data#38401)
## Description of the change

This change improves the performance of sandbox pipeline image builds by
excluding a large set of files from the set of code files that get
copied to our Docker image.

[Here](https://console.cloud.google.com/cloud-build/builds;region=global/a68c32e0-3fd1-41f8-bbf8-3aa18531e33e?authuser=0&organizationId=448885369991&project=recidiviz-staging)
is a build that I ran with this change. It took 2:13 total. Copying
files took 26 seconds.

[Here](https://console.cloud.google.com/cloud-build/builds;region=global/afc8ba94-993a-484b-8c42-34197259f3b5?authuser=0&organizationId=448885369991&project=recidiviz-staging)
is a build that ran at a similar time without this change. It took 3:08.
Copying files took 49 seconds.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

N/A

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3bce329096fcf82c3b64f6198935cbe615dc3c3e