[Justice Counts] Add `before_commit` Trigger for Reporting Agency Cleanup on Deletion (Recidiviz/recidiviz-data#38286)
This PR introduces a SQLAlchemy event trigger to handle the cleanup of
`reporting_agency_id` in `MetricSetting` when a vendor or agency is
deleted. Previously, deleting a vendor or agency could lead to foreign
key constraint violations because MetricSetting stored the
reporting_agency_id in a JSONB field, and the deletion was not
automatically propagated.

To resolve this, a `before_commit` trigger ensures that any
MetricSetting referencing the deleted vendor or agency is updated to
nullify reporting_agency_id before the transaction is committed.

SQLAlchemy allows us to listen for changes in the database using event
listeners. In this case, I used `before_commit`, which fires before the
transaction is committed. This ensures that any cleanup happens before
deletion, preventing constraint issues.

If we continue running into issues with reporting agency deletions, I
will consider moving reporting_agency_id to a separate database column
instead of storing it in JSONB. This would make deletion cascades and
updates more efficient and avoid needing manual cleanup logic. However,
for now, this approach should work without requiring a migration.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5e2bd42ebcd377eeff30807b54b1fa1fcdd3f6a5