[ingest] adds `sandbox_raw_dataset_prefix` option to `ingest_stability_check` (Recidiviz/recidiviz-data#38436)
## Description of the change

adds the `sandbox_raw_dataset_prefix` option to `ingest_stability_check`
in order to allow IEs to run the `raw_data` primary key check based on a
sandbox raw data import. if the prefix is specified, will only run the
checks on the tables in the sandbox

note: this only makes sense to do for the `raw_data` run right now (i
think). i tried to trace how we would provide a dataset override in
`DirectIngestViewQueryBuilder` but seemed like there wasn't one natively
supported (either i'm missing it or it was a decision to do it this
way?)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#38090

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1538c89700e7eb36a2c0b632272c18b206fcb080