[Sentencing][US_ND] Include counties in `us_nd_location_metadata` and includes in the sentencing counties export (Recidiviz/recidiviz-data#38465)
## Description of the change
The county and district data for ND has finally made its way to staging,
so now we can include it in `location_metadata` and the sentencing
metric export for counties and districts.

I generated the changed views locally + tested the resulting metric
export against the expected import schema and everything looks correct!

Also RIP the number of changes to the view graph.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Fixes https://github.com/Recidiviz/recidiviz-data/issues/38109

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 835c9a4c237c8323168f0d141b96e4e9c5e35303