[Testing] Define new fixture paths via functions (Recidiviz/recidiviz-data#38388)
## Description of the change

This PR defines the three options to get an ingest fixture path:
- The function for enum parsing test fixtures
- The function for ingest mapping test _output_
- A function that returns a fixture path for a BigQuery address (all
remaining fixtures)

Note that the v1 fixtures are currently untouched, except for the enum
parsing test fixtures which are now in their final resting place.

I've also migrated our check for unused ingest view results fixtures, so
that the structure test is ready to go for the migration in
https://github.com/Recidiviz/recidiviz-data/issues/38322

I've technically dropped the check that an existing ingest view test has
the required fixtures, but those tests will fail on their own if that is
the case so thought that was an OK change.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| **Type: Feature** | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#38323

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: bbc7e49214a4fa18e3fc54e6d3dc532b6e45dc30