[US_TX][Workflows] TX ARS ineligible offense criterion  (Recidiviz/recidiviz-data#38378)
## Description of the change

First pass at TX ARS ineligible offense criterion, which creates spans
of time when clients do not have any current or prior convictions for
the offenses listed
[here](https://statutes.capitol.texas.gov/Docs/CR/htm/CR.42A.htmRecidiviz/recidiviz-data#42A.054).

Notes for @pepinse:
- Some offenses are commented out where we don't have enough information
to check them (e.g. we do not have information about felonies,
subsections, etc.) so I'll let you decide what to do with those (in PA
we have some similar cases and we ended up putting a case note in the
sidebar that says like "hey they have this charge but it only makes them
ineligible if the victim was a child, check criminal history for that").
Or if we can get more data from TX that would work too
- This was created off of an early sandbox version of `state_charge_v2`
with a subset of data, so you might need to refine this logic once we
get the full historical data. I would also double check with Santy that
we are only getting convictions (no dismissed/withdrawn cases or other
statuses) or filter for status
- The bigquery test is failing because there's no TX eligibility span to
base it off of yet, so you'll need to merge in your other PR and add
this criterion to the eligibility span before merging this in. I'm not
sure how merging in someone else's PR works so if you need to just copy
this code into your own PR feel free!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#38045

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Sophie Pepin <50818505+pepinse@users.noreply.github.com>
GitOrigin-RevId: 11a134ccb609f428d969ba22a6db6cf145a78757