[US_AR] Address COMPLETED/VACATED sentence statuses followed by other statuses (Recidiviz/recidiviz-data#38501)
## Description of the change

Ingest pipelines are failing because of sentence status snapshots with a
status of COMPLETED or VACATED despite being followed by other statuses
for the same sentence. This change fixes the issue by:
- setting status to INTERNAL_UNKNOWN for snapshots with a
COMPLETED/VACATED status, if the snapshot is followed by other statuses
that aren't COMPLETED or VACATED
- removing COMPLETED/VACATED snapshots entirely if they're immediately
preceded by another COMPLETED/VACATED status

See Recidiviz/recidiviz-data#38496 for follow-up work on a long term solution.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 30700583a6b81f6e4836b0c7972a9a1227cbf093