[raw data] Refactors in deprecate_raw_data (Recidiviz/recidiviz-data#38508)
## Description of the change

we recently saw
[issue](https://github.com/Recidiviz/recidiviz-data/issues/38234) where
`deprecate_raw_data` script was failing when supplied a list of file
tags and there were some directories that all file tags were not present
in.

this is because we determine the superset of subdirectories for all file
tags and try to mv in all of those, knowing that some might fail. in the
original implementation we passed `allow_empty=True` to `gsutil_mv` so
if we received a message that `No URLs matched: ...` we just ignored it
and kept going.
BUT we were running into
[issue](https://github.com/Recidiviz/recidiviz-data/issues/25501) where
the script would say its deprecating files but not actually deprecate
anything if we supplied a `file_tag_filter` that was a substring match
but not an exact match for a file tag, and because of the `allow_empty`
flag we were just failing silently. so we added a
[change](https://github.com/Recidiviz/recidiviz-data/pull/25852) to fail
loudly which makes sense but negates the functionality of the
`allow_empty` flag because we just ignore the flag and always throw the
exception.

the original issue we were facing with silent failures was due to us
looking for subdirectories to operate on searching with suffix match
`*{file_tag_filter}*` then doing the actual move operation using suffix
match `*_{file_tag_filter}[.]*`

SO this pr:
- stores the controller's `file_tag_filters` and `file_tag_regex` as the
actual filter/regex that will be applied so we use the same for both
searching for subdirs and performing the operation
- revives the functionality of the `allow_empty` flag to allow us to
ignore exceptions caused by no matching files in a `gsutil_cp/mv`
operation
- adds an if clause to return early if there are no subdirs with
matching files in them so i can close out this old
[issue](https://github.com/Recidiviz/recidiviz-data/issues/18842)
- adds some docstrings and refactors for readiblity

i tested this locally and also opened this
[issue](https://github.com/Recidiviz/recidiviz-data/issues/38507) to
write UTs for the controller in a follow up

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#38234

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 50d19d48c1e3ad55dab47870220320da7912e603