[US_TX] Removing officer from supervision period when not present in staff data (Recidiviz/recidiviz-data#38517)
## Description of the change

There are times we are given supervision periods which contain
supervision officer IDs for which there is no existing corresponding
staff in the data. In these cases, we remove the officer from the
period. Doing this so that the pipeline can succeed.

Also have a ticket here to investigate this further here: Recidiviz/recidiviz-data#38519

[SUCCESSFUL SANDBOX HERE

](https://console.cloud.google.com/bigquery?ws=!1m4!1m3!3m2!1srecidiviz-staging!2ssanty_test_us_tx_state)##
Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#38516

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1984497becaa96a52cd32441c53e23c9a9e2855a